requirements.txt
# Mental Health ML Project - Requirements
# DLBDSMLUSL01 – Machine Learning – Unsupervised Learning

# Core Data Science
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
scipy>=1.11.0

# Dimensionality Reduction
umap-learn>=0.5.3

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Jupyter
jupyter>=1.0.0
notebook>=7.0.0
ipywidgets>=8.0.0

# Utilities
tqdm>=4.65.0

# Optional: For better performance
joblib>=1.3.0

